Tag-it AI/requirements.txt
flask
flask-cors
openai==1.17.0
python-dotenv
httpx
langdetect
PyPDF2
python-docx